new updates to p1

+++ p1/Models/Tickets.cs
@@ -7,7 +7,7 @@ namespace Models
 {
     public class Tickets
     {
-        public Tickets(decimal ticketamount, int ticketstatus, string notes, int user_fk, Guid requestid)
+        public Tickets(float ticketamount, int ticketstatus, string notes, int user_fk, Guid requestid)
         {
             TicketAmount = ticketamount;
             TicketStatus = ticketstatus;
@@ -17,7 +17,7 @@ namespace Models
         }
         
         //methods that coincide with the db tickets table's attributes.
-        public decimal TicketAmount {get; set;}
+        public float TicketAmount {get; set;}
         public int TicketStatus { get; set;}
         public string Notes { get; set;}
         public int User_Fk { get; set;}
